remade the bot into 1 class with 3 methods
