+++ .gitignore
@@ -1,1 +1,2 @@
 .env
+lib/v2.rb
